+++ activity/src/main/java/pt/ulisboa/tecnico/softeng/activity/services/remote/dataobjects/RestBankOperationData.java
@@ -10,7 +10,7 @@ public class RestBankOperationData {
 	private String reference;
 	private String type;
 	private String iban;
-	private Double value;
+	private long value;
 	@JsonDeserialize(using = DateTimeDeserializer.class)
 	@JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss.SSS")
 	private DateTime time;
@@ -20,7 +20,7 @@ public class RestBankOperationData {
 	public RestBankOperationData() {
 	}
 
-	public RestBankOperationData(String iban, double value, String transactionSource, String transactionReference) {
+	public RestBankOperationData(String iban, long value, String transactionSource, String transactionReference) {
 		this.iban = iban;
 		this.value = value;
 		this.transactionSource = transactionSource;
@@ -51,11 +51,11 @@ public class RestBankOperationData {
 		this.iban = iban;
 	}
 
-	public Double getValue() {
+	public long getValue() {
 		return this.value;
 	}
 
-	public void setValue(Double value) {
+	public void setValue(long value) {
 		this.value = value;
 	}
